Add a Tuple::makeTuple function to easily construct a tuple. Update Tuple to allow all types to be passed via .append() so they can be used with makeTuple.

+++ fdbclient/FileBackupAgent.actor.cpp
@@ -203,7 +203,7 @@ public:
 			return Tuple()
 			    .append(version)
 			    .append(StringRef(fileName))
-			    .append(isRange)
+			    .append((int)isRange)
 			    .append(fileSize)
 			    .append(blockSize)
 			    .append(endVersion);
